+++ Framework/CloudSDK/Changelog.md
@@ -2,6 +2,7 @@
 
 ## Current master
 
+- Implement get location based app by id endpoint (#23)
 - [Breaking] Rewrite cloud sdk from scratch & Include pace endpoints (#19)
 - Implement first iteration (#1)
 - Prepare for first release (#2)